로그인 실패 메세지 출력 & 완료된 프로젝트 목록 출력
- 서버 연결 실패시 SnackBar가 안뜨던 오류를 수정함
- 참가중인 프로젝트 버튼 추가
    - 누르면 ProjectModel에 저장된 _endedProjects(완료된 프로젝트 목록)을 출력
    - _projects <-> _endedProjects에 따라 텍스트 버튼의 텍스트도 바뀌도록 구현
- 작업 파일명 work_with_demo로 변경
