+++ work_with_demo/README.md
@@ -1,6 +1,6 @@
-# login_demo
+# work_with_demo
 
-LogIn Demonstration.
+Work With demo version for DX sprint 2023
 
 # 2023/07/13
 - 로그인 화면 작업
@@ -30,5 +30,12 @@ LogIn Demonstration.
 # 2023/07/19
 - test_app : Flutter을 통한 Web socket 통신 테스트
 - socket & login_demo :
-  - socket : 8081 -> web socket 로컬 서버, 8090 -> 로그인 HTTP 로컬 서버
-  - login_demo : 로컬 서버 8090과의 통신으로 수신-응답이 제대로 이루어졌을때 ViewProjectPage로 넘어가도록 수정
+    - socket : 8081 -> web socket 로컬 서버, 8090 -> 로그인 HTTP 로컬 서버
+    - login_demo : 로컬 서버 8090과의 통신으로 수신-응답이 제대로 이루어졌을때 ViewProjectPage로 넘어가도록 수정
+
+# 2023/07/20
+- 서버 연결 실패시 SnackBar가 안뜨던 오류를 수정함
+- 참가중인 프로젝트 버튼 추가
+    - 누르면 ProjectModel에 저장된 _endedProjects(완료된 프로젝트 목록)을 출력
+    - _projects <-> _endedProjects에 따라 텍스트 버튼의 텍스트도 바뀌도록 구현
+- 작업 파일명 work_with_demo로 변경
