+++ work_with_demo/pubspec.yaml
@@ -1,5 +1,5 @@
-name: login_demo
-description: LogIn Demonstration.
+name: work_with_demo
+description: Work With demo version
 
 publish_to: 'none'
