+++ work_with_demo/web/index.html
@@ -18,18 +18,18 @@
 
   <meta charset="UTF-8">
   <meta content="IE=Edge" http-equiv="X-UA-Compatible">
-  <meta name="description" content="LogIn Demonstration.">
+  <meta name="description" content="Work With demo version">
 
   <!-- iOS meta tags & icons -->
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="login_demo">
+  <meta name="apple-mobile-web-app-title" content="work_with_demo">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>login_demo</title>
+  <title>work_with_demo</title>
   <link rel="manifest" href="manifest.json">
 
   <script>
